requirements.txt
fastapi
uvicorn
python-multipart
PyPDF2
python-docx
openai
faiss-cpu
python-dotenv
fastapi-jwt-auth